ConfigurationService internal implementation. Prevent site deletion from stopping on invalidate-cache error
