+++ src/main/java/org/craftercms/studio/impl/v1/service/site/SiteServiceImpl.java
@@ -162,6 +162,7 @@ public class SiteServiceImpl implements SiteService, ApplicationContextAware {
     protected SitesService sitesServiceInternal;
     protected AuditServiceInternal auditServiceInternal;
     protected ConfigurationService configurationService;
+    protected ConfigurationService configurationServiceInternal;
     protected ItemServiceInternal itemServiceInternal;
     protected WorkflowServiceInternal workflowServiceInternal;
     protected ApplicationContext applicationContext;
@@ -815,7 +816,6 @@ public class SiteServiceImpl implements SiteService, ApplicationContextAware {
 
         try {
             logger.debug("Delete the Deployer targets for site '{}'", siteId);
-
             deployer.deleteTargets(siteId);
         } catch (Exception e) {
             success = false;
@@ -824,7 +824,6 @@ public class SiteServiceImpl implements SiteService, ApplicationContextAware {
 
         try {
             logger.debug("Destroy the preview context for site '{}'", siteId);
-
             success = success && destroySitePreviewContext(siteId);
         } catch (Exception e) {
             success = false;
@@ -833,16 +832,17 @@ public class SiteServiceImpl implements SiteService, ApplicationContextAware {
 
         try {
             logger.debug("Delete the git repo for site '{}'", siteId);
-
             contentRepository.deleteSite(siteId);
         } catch (Exception e) {
             success = false;
             logger.error("Failed to delete the repository for site '{}'", siteId, e);
         }
 
-        // clear cache
-        configurationService.invalidateConfiguration(siteId);
-
+        try {
+            configurationServiceInternal.invalidateConfiguration(siteId);
+        } catch (Exception e) {
+            logger.error("Failed to invalidate the configuration for site '{}'", siteId, e);
+        }
         try {
             // delete database records
             logger.debug("Delete the database records for site '{}'", siteId);
@@ -1757,6 +1757,10 @@ public class SiteServiceImpl implements SiteService, ApplicationContextAware {
 
     public void setConfigurationService(ConfigurationService configurationService) {
         this.configurationService = configurationService;
+    }
+
+    public void setConfigurationServiceInternal(final ConfigurationService configurationServiceInternal) {
+        this.configurationServiceInternal = configurationServiceInternal;
     }
 
     public void setContentRepositoryV2(ContentRepository contentRepositoryV2) {
